# api/openapi/discount.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Discount service
  description: One stop solution for loyality program
  contact:
    name: Dmitry Makarenko
servers:
  - url: http://{hostname}/api
    variables:
      hostname:
        default: localhost
security:
  - bearerAuth: [ ]
paths:
  /brand/{brandID}/code/generate:
    post:
      operationId: generateCode
      parameters:
        - in: path
          name: brandID
          example: my_super_popular_brand
          schema:
            type: string
          required: true
          description: ID of the brand
      requestBody:
        description: Parameters of generated code
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDiscountCode'
      responses:
        '204':
          description: Succesfully generate discount code
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /brand/{brandID}/code/fetch:
    post:
      operationId: fetchCode
      parameters:
        - in: path
          name: brandID
          example: my_super_popular_brand
          schema:
            type: string
          required: true
          description: ID of the brand
      responses:
        '200':
          description: Successfuly fetch code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCode'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    DiscountCode:
      type: object
      required: [ code, brand ]
      properties:
        code:
          type: string
          example: "123e4567-e89b-12d3-a456-426614174000"
        brand:
          type: string
          example: my_super_popular_brand
    GenerateDiscountCode:
      type: object
      required: [ amount ]
      properties:
        amount:
          type: integer
          minimum: 1
          maximum: 200
          default: 1
          description: The number of codes to generate
    Error:
      type: object
      required:
        - slug
        - message
      properties:
        slug:
          type: string
        message:
          type: string