# internal/discount/ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package ports

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// DiscountCode defines model for DiscountCode.
type DiscountCode struct {
	Brand string `json:"brand"`
	Code  string `json:"code"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
	Slug    string `json:"slug"`
}

// GenerateDiscountCode defines model for GenerateDiscountCode.
type GenerateDiscountCode struct {
	// The number of codes to generate
	Amount int `json:"amount"`
}

// GenerateCodeJSONBody defines parameters for GenerateCode.
type GenerateCodeJSONBody GenerateDiscountCode

// GenerateCodeJSONRequestBody defines body for GenerateCode for application/json ContentType.
type GenerateCodeJSONRequestBody GenerateCodeJSONBody